testauth/requirements.txt
django == 1.9.5
djangorestframework == 3.3.3
psycopg2 == 2.6.1
dj-database-url==0.4.0
gunicorn==19.4.5
whitenoise==2.0.6